Make Ban Checking Constant Time
Ban checking is now constant time. This makes ban checking for a small
banlist slower by a factor of 10. However, more importantly, ban
checking now takes the same amount of time regardless of how many bans
you have stored.

 Original Pyspades

10 8.670362993143498e-06
100 5.076720300712623e-05
1000 0.0004664193200005684
10000 0.004667077518010047

 New Constant Time

10 9.937950099993032e-05
100 0.00010137572599342092
1000 0.00010073935400578194
10000 0.00010321205800573807

It would be very easy to make this faster even for lower values.
IPv4Address objects are heavy to create, and we create a bunch of them.
They could be created only when necessary for output, and a faster (ip:
int, mask: int) tuple could be used internally.

However, this commit solves the problem aloha is having and I have to go
now.