badmin.py: Check also for rifle percentage
Without this change whenever you would check for a player that is using rifle with /investigate
the command would fail because it did not return anything thus being type None which would
cause issues with round().

semi_hits doesn't have to exist as we have rifle_hits which is what we should use for checking
rifle hits and counts.

With this fix badmin finally checks also for players using rifle and doesn't fail on that.