ci: only trigger a wheel build on release and manual dispatch
The "create" events triggers on tag or branch creation, which is not super
useful, and also creates duplicate runs on release (since we also create a
tag there).

The broad "release" event would trigger a build on every release-related
event, so whenever a release is published, unpublished, created, edited, ...
We don't want that.