fix(master): catch dns lookup failures inside `add_client()`
We were already catching DNS lookup failures, but only on first connection,
not on subsequent reconnects.

This can happen if the server starts with a working internet connection, but
then loses connection for over 60 seconds (the reconnect interval). Since the
reconnect call is not using try-except, we get an uncaught error.

This commit moves the try-except inside `add_client` directly so we catch it
on first connect and on any reconnect.